[tmp] committing for test purposes

+++ proxy/vless/inbound/inbound.go
@@ -593,8 +593,8 @@ func (h *Handler) Process(ctx context.Context, network net.Network, connection s
 		remoteAddr,
 		strings.ToLower(strings.TrimSpace(request.Destination().Address.String())),
 		uint16(request.Destination().Port),
-		0,
-		0,
+		uint64(0),
+		uint64(0),
 		time.Since(now),
 	)
